requirements.txt
Babel==2.8.0
dj-database-url==0.5.0
Django==2.2.11
django-extra-views==0.11.0
django-filter==2.2.0
django-haystack==2.8.1
django-oscar==2.0.4
django-phonenumber-field==2.0.1
django-tables2==1.21.2
django-treebeard==4.3.1
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
docopt==0.6.2
factory-boy==2.12.0
Faker==4.0.3
gitignore==0.0.8
gunicorn==19.9.0
Markdown==3.2.1
phonenumbers==8.12.1
Pillow==7.1.1
pkg-resources==0.0.0
psycopg2==2.8.5
purl==1.5
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
static3==0.7.0
text-unidecode==1.3
whitenoise==5.0.1
